Moved FetchMoviesTask to its own file. Using FrameLayout instead of Relative in main fragment. Updated gradle version.
