+++ app/src/main/res/layout/fragment_main.xml
@@ -1,4 +1,4 @@
-<RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
+<FrameLayout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
@@ -16,4 +16,4 @@
         android:stretchMode="columnWidth"
         android:verticalSpacing="8dp" />
 
-</RelativeLayout>
+</FrameLayout>
